fix build issues

+++ beacon-chain/sync/initial-sync/blocks_queue.go
@@ -84,7 +84,7 @@ func newBlocksQueue(ctx context.Context, cfg *blocksQueueConfig) *blocksQueue {
 		highestExpectedSlot: highestExpectedSlot,
 		blocksFetcher:       blocksFetcher,
 		headFetcher:         cfg.headFetcher,
-		fetchedBlocks:       make(chan *eth.SignedBeaconBlock, allowedBlocksPerSecond),
+		fetchedBlocks:       make(chan *eth.SignedBeaconBlock, int64(allowedBlocksPerSecond)),
 		quit:                make(chan struct{}),
 	}
