+++ beacon-chain/sync/initial-sync/round_robin.go
@@ -91,7 +91,7 @@ func (s *Service) roundRobinSync(genesis time.Time) error {
 	for head := helpers.SlotsSince(genesis); s.chain.HeadSlot() < head; {
 		req := &p2ppb.BeaconBlocksByRangeRequest{
 			StartSlot: s.chain.HeadSlot() + 1,
-			Count:     mathutil.Min(helpers.SlotsSince(genesis)-s.chain.HeadSlot()+1, allowedBlocksPerSecond),
+			Count:     mathutil.Min(helpers.SlotsSince(genesis)-s.chain.HeadSlot()+1, uint64(allowedBlocksPerSecond)),
 			Step:      1,
 		}
